backend/repos/static/dbcr/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-shell==1.0
dnspython==1.16.0
git-filter-repo==2.24.0
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
mysql-connector-python==8.0.19
protobuf==3.6.1
python-gitlab==1.13.0
requests==2.22.0
six==1.13.0
smmap2==2.0.5
urllib3==1.25.7
